+++ CHANGELOG.md
@@ -34,6 +34,7 @@ This change was introduced in [pull request #1796: Hide character count message 
 - [#1795: Add all page template variables from GOV.UK Frontend](https://github.com/nhsuk/nhsuk-frontend/pull/1795)
 - [#1797: Ensure components pass label.id down to the label](https://github.com/nhsuk/nhsuk-frontend/pull/1797)
 - [#1807: Fix summary list cells becoming vertically misaligned when a multi-line inline-block element is present](https://github.com/nhsuk/nhsuk-frontend/pull/1807)
+- [#1829: Always validate component config schema](https://github.com/nhsuk/nhsuk-frontend/pull/1829)
 
 ## 10.3.1 - 19 January 2026